JSSpeccy.js runtime specifies ROMs symbolically to the worker.js
The Core now just has 3 page slots for ROMs: ROM0, ROM1 & BetaDisk,
and the worker.js loads in the correct ROMs for the machine when
setting the machine type.

The worker.js keeps track of symbolic ROM names, and copies the
ROM images into the WASM core when the `setMachine` message is
received (or when we load a snapshot).

So now jsspeccy.js sends a multiple 'loadRom' messages at startup,
with a symbolic ROM image name (instead of sending 'loadMemory'
messages, and knowing which ROM image goes into which page slot).